Unify logging integration tests (#1017)
* Unify max samples stored tests

* Unify CSP/HSM disabling tests

* Unify local decoration tests

* Use log framework enum in other tests

* Unify metrics and forwarding tests
