--- tests/Agent/IntegrationTests/IntegrationTests/Logging/Log4netJsonLayoutDecorationTests.cs
@@ -1,176 +1,0 @@
-﻿// Copyright 2020 New Relic, Inc. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-using System.Text.RegularExpressions;
-using MultiFunctionApplicationHelpers;
-using NewRelic.Agent.IntegrationTestHelpers;
-using Xunit;
-using Xunit.Abstractions;
-
-namespace NewRelic.Agent.IntegrationTests.Logging
-{
-    public abstract class Log4netJsonLayoutDecorationTestsBase<TFixture> : NewRelicIntegrationTest<TFixture>
-        where TFixture : ConsoleDynamicMethodFixture
-    {
-        private readonly TFixture _fixture;
-        private readonly bool _decorationEnabled;
-
-        public Log4netJsonLayoutDecorationTestsBase(TFixture fixture, ITestOutputHelper output, bool decorationEnabled) : base(fixture)
-        {
-            _decorationEnabled = decorationEnabled;
-            _fixture = fixture;
-            _fixture.SetTimeout(System.TimeSpan.FromMinutes(2));
-            _fixture.TestLogger = output;
-
-            _fixture.AddCommand($"LoggingTester SetFramework log4net");
-            _fixture.AddCommand($"LoggingTester ConfigureJsonLayoutAppenderForDecoration");
-            _fixture.AddCommand($"LoggingTester CreateSingleLogMessage DecorateMe DEBUG");
-
-            _fixture.Actions
-            (
-                setupConfiguration: () =>
-                {
-                    var configModifier = new NewRelicConfigModifier(fixture.DestinationNewRelicConfigFilePath);
-
-                    configModifier
-                    .EnableApplicationLogging()
-                    .EnableLogDecoration(_decorationEnabled)
-                    .EnableDistributedTrace()
-                    .SetLogLevel("debug");
-                }
-            );
-
-            _fixture.Initialize();
-        }
-
-        [Fact]
-        public void LogIsDecorated()
-        {
-            // Sample decorated data we are looking for:
-            // "NR-LINKING|MjczMDcwfEFQTXxBUFBMSUNBVElPTnwxODQyMg|blah.hsd1.ca.comcast.net|45f120972d61834b96fb890d2a8f97e7|840d9a82e8bc18a8|"
-            var regex = new Regex(@"NR-LINKING\|[a-zA-Z0-9]*\|[a-zA-Z0-9._-]*\|[a-zA-Z0-9]*\|[a-zA-Z0-9]*\|");
-            if (_decorationEnabled)
-            {
-                Assert.Matches(regex, _fixture.RemoteApplication.CapturedOutput.StandardOutput);
-            }
-            else
-            {
-                Assert.DoesNotMatch(regex, _fixture.RemoteApplication.CapturedOutput.StandardOutput);
-            }
-        }
-    }
-
-    #region Enabled Tests
-    [NetFrameworkTest]
-    public class Log4netJsonLayoutDecorationEnabledTestsFWLatestTests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureFWLatest>
-    {
-        public Log4netJsonLayoutDecorationEnabledTestsFWLatestTests(ConsoleDynamicMethodFixtureFWLatest fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetFrameworkTest]
-    public class Log4netJsonLayoutDecorationEnabledTestsFW471Tests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureFW471>
-    {
-        public Log4netJsonLayoutDecorationEnabledTestsFW471Tests(ConsoleDynamicMethodFixtureFW471 fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetCoreTest]
-    public class Log4netJsonLayoutDecorationEnabledTestsNetCoreLatestTests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureCoreLatest>
-    {
-        public Log4netJsonLayoutDecorationEnabledTestsNetCoreLatestTests(ConsoleDynamicMethodFixtureCoreLatest fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetCoreTest]
-    public class Log4netJsonLayoutDecorationEnabledTestsNetCore50Tests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureCore50>
-    {
-        public Log4netJsonLayoutDecorationEnabledTestsNetCore50Tests(ConsoleDynamicMethodFixtureCore50 fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetCoreTest]
-    public class Log4netJsonLayoutDecorationEnabledTestsNetCore31Tests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureCore31>
-    {
-        public Log4netJsonLayoutDecorationEnabledTestsNetCore31Tests(ConsoleDynamicMethodFixtureCore31 fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetCoreTest]
-    public class Log4netJsonLayoutDecorationEnabledTestsNetCore22Tests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureCore22>
-    {
-        public Log4netJsonLayoutDecorationEnabledTestsNetCore22Tests(ConsoleDynamicMethodFixtureCore22 fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-    #endregion
-
-    #region Disabled Tests
-    [NetFrameworkTest]
-    public class Log4netJsonLayoutDecorationDisabledTestsFWLatestTests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureFWLatest>
-    {
-        public Log4netJsonLayoutDecorationDisabledTestsFWLatestTests(ConsoleDynamicMethodFixtureFWLatest fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetFrameworkTest]
-    public class Log4netJsonLayoutDecorationDisabledTestsFW471Tests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureFW471>
-    {
-        public Log4netJsonLayoutDecorationDisabledTestsFW471Tests(ConsoleDynamicMethodFixtureFW471 fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetCoreTest]
-    public class Log4netJsonLayoutDecorationDisabledTestsNetCoreLatestTests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureCoreLatest>
-    {
-        public Log4netJsonLayoutDecorationDisabledTestsNetCoreLatestTests(ConsoleDynamicMethodFixtureCoreLatest fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetCoreTest]
-    public class Log4netJsonLayoutDecorationDisabledTestsNetCore50Tests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureCore50>
-    {
-        public Log4netJsonLayoutDecorationDisabledTestsNetCore50Tests(ConsoleDynamicMethodFixtureCore50 fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetCoreTest]
-    public class Log4netJsonLayoutDecorationDisabledTestsNetCore31Tests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureCore31>
-    {
-        public Log4netJsonLayoutDecorationDisabledTestsNetCore31Tests(ConsoleDynamicMethodFixtureCore31 fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-
-    [NetCoreTest]
-    public class Log4netJsonLayoutDecorationDisabledTestsNetCore22Tests : Log4netJsonLayoutDecorationTestsBase<ConsoleDynamicMethodFixtureCore22>
-    {
-        public Log4netJsonLayoutDecorationDisabledTestsNetCore22Tests(ConsoleDynamicMethodFixtureCore22 fixture, ITestOutputHelper output)
-            : base(fixture, output, true)
-        {
-        }
-    }
-    #endregion
-
-
-}
